requirements.txt
antiorm==1.2.1
asgiref==3.3.1
db==0.1.1
db-sqlite3==0.0.1
Django==3.1.3
django-filter==2.4.0
mysqlclient==2.0.1
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
